+++ sdks/python/hatchet_sdk/clients/rest/models/tenant.py
@@ -23,6 +23,7 @@ from pydantic import BaseModel, ConfigDict, Field, StrictBool, StrictStr
 from typing_extensions import Self
 
 from hatchet_sdk.clients.rest.models.api_resource_meta import APIResourceMeta
+from hatchet_sdk.clients.rest.models.tenant_environment import TenantEnvironment
 from hatchet_sdk.clients.rest.models.tenant_ui_version import TenantUIVersion
 from hatchet_sdk.clients.rest.models.tenant_version import TenantVersion
 
@@ -49,6 +50,9 @@ class Tenant(BaseModel):
     ui_version: Optional[TenantUIVersion] = Field(
         default=None, description="The UI of the tenant.", alias="uiVersion"
     )
+    environment: Optional[TenantEnvironment] = Field(
+        default=None, description="The environment type of the tenant."
+    )
     __properties: ClassVar[List[str]] = [
         "metadata",
         "name",
@@ -57,6 +61,7 @@ class Tenant(BaseModel):
         "alertMemberEmails",
         "version",
         "uiVersion",
+        "environment",
     ]
 
     model_config = ConfigDict(
@@ -123,6 +128,7 @@ class Tenant(BaseModel):
                 "alertMemberEmails": obj.get("alertMemberEmails"),
                 "version": obj.get("version"),
                 "uiVersion": obj.get("uiVersion"),
+                "environment": obj.get("environment"),
             }
         )
         return _obj