+++ sdks/python/hatchet_sdk/worker/runner/utils/capture_logs.py
@@ -1,13 +1,18 @@
 import asyncio
+import copy
 import functools
 import logging
 from collections.abc import Awaitable, Callable
+from datetime import datetime, timedelta, timezone
 from io import StringIO
 from typing import Literal, ParamSpec, TypeVar
+from uuid import UUID, uuid4
 
+import grpc
 from pydantic import BaseModel, Field
 
-from hatchet_sdk.clients.events import EventClient
+from hatchet_sdk.clients.events import BulkLogPushRequest, EventClient
+from hatchet_sdk.config import ClientConfig
 from hatchet_sdk.logger import logger
 from hatchet_sdk.runnables.contextvars import (
     ctx_action_key,
@@ -70,9 +75,31 @@ def copy_context_vars(
 
 
 class LogRecord(BaseModel):
+    id: UUID = Field(default_factory=uuid4)
     message: str
-    step_run_id: str
+    task_run_external_id: str
     level: LogLevel
+    timestamp: datetime
+
+    def __hash__(self) -> int:
+        return hash(self.id)
+
+
+class LogBuffer(BaseModel):
+    records: set[LogRecord] = Field(default_factory=set)
+
+    def should_clear(self, config: ClientConfig) -> bool:
+        if not self.records:
+            return False
+
+        if len(self.records) >= config.log_buffer_size:
+            return True
+
+        oldest_ts = min(r.timestamp for r in self.records)
+
+        return datetime.now(timezone.utc) - oldest_ts > timedelta(
+            seconds=config.log_flush_interval_seconds
+        )
 
 
 class AsyncLogSender:
@@ -81,23 +108,78 @@ class AsyncLogSender:
         self.q = asyncio.Queue[LogRecord | STOP_LOOP_TYPE](
             maxsize=event_client.client_config.log_queue_size
         )
+        self.buffer = LogBuffer()
+
+        ## flag to see if the engine supports bulk log push
+        ## defaults to true, but reset to false if we get an
+        ## `UNIMPLEMENTED` or `NOT_FOUND` error back
+        self.bulk_log_push_enabled = True
+
+    async def flush(self) -> None:
+        ## IMPORTANT: Need a deepcopy here to act as a snapshot of the
+        ## current buffer, since new logs might be added while we're processing
+        ## the current buffer
+        records = copy.deepcopy(self.buffer.records)
+
+        requests = [
+            BulkLogPushRequest(
+                message=r.message,
+                task_run_external_id=r.task_run_external_id,
+                level=r.level,
+            )
+            for r in records
+        ]
+
+        try:
+            await asyncio.to_thread(self.event_client.bulk_log, requests)
+            self.buffer.records -= records
+        except grpc.RpcError as e:
+            if e.code() in [grpc.StatusCode.UNIMPLEMENTED, grpc.StatusCode.NOT_FOUND]:
+                self.bulk_log_push_enabled = False
+                for record in records:
+                    await self.push_single(record)
+                    self.buffer.records.remove(record)
+
+        except Exception:
+            logger.exception("failed to send log to Hatchet")
+
+    async def push_single(self, record: LogRecord) -> None:
+        try:
+            await asyncio.to_thread(
+                self.event_client.log,
+                message=record.message,
+                step_run_id=record.task_run_external_id,
+                level=record.level,
+            )
+        except Exception:
+            logger.exception("failed to send log to Hatchet")
+
+    async def get_log(self) -> LogRecord | STOP_LOOP_TYPE | None:
+        try:
+            async with asyncio.timeout(
+                self.event_client.client_config.log_flush_interval_seconds
+            ):
+                return await self.q.get()
+        except TimeoutError:
+            return None
 
     async def consume(self) -> None:
         while True:
-            record = await self.q.get()
+            record = await self.get_log()
 
             if record == STOP_LOOP:
+                await self.flush()
                 break
 
-            try:
-                await asyncio.to_thread(
-                    self.event_client.log,
-                    message=record.message,
-                    step_run_id=record.step_run_id,
-                    level=record.level,
-                )
-            except Exception:
-                logger.exception("failed to send log to Hatchet")
+            if self.bulk_log_push_enabled:
+                if record:
+                    self.buffer.records.add(record)
+
+                if self.buffer.should_clear(self.event_client.client_config):
+                    await self.flush()
+            else:
+                if record:
+                    await self.push_single(record)
 
     def publish(self, record: LogRecord | STOP_LOOP_TYPE) -> None:
         try:
@@ -124,8 +206,9 @@ class LogForwardingHandler(logging.StreamHandler):  # type: ignore[type-arg]
         self.log_sender.publish(
             LogRecord(
                 message=log_entry,
-                step_run_id=step_run_id,
+                task_run_external_id=step_run_id,
                 level=LogLevel.from_levelname(record.levelname),
+                timestamp=datetime.now(timezone.utc),
             )
         )